+++ src/action/transfer/send/order.ts
@@ -7,7 +7,6 @@ import {
 import * as OrderFactory from '../../../order';
 import { IGood, IOwnershipInfo } from '../../../ownershipInfo';
 import * as SendActionFactory from '../send';
-import { IAttributes as ISendEmailMessageActionAttributes } from './message/email';
 
 export type IAgent = IParticipantAsPerson | IParticipantAsProject | IParticipantAsWebApplication;
 export type IRecipient = IParticipantAsWebApplication | IParticipantAsPerson | IParticipantAsCustomer;
@@ -24,20 +23,19 @@ export type IObject = OrderFactory.ISimpleOrder & {
  * 注文配送結果としての所有権
  */
 export type IResult = IOwnershipInfo<IGood>[];
-export type ISendEmailMessage = Pick<ISendEmailMessageActionAttributes, 'object'>;
-export interface IPotentialActions {
-    // moneyTransfer?: IMoneyTransferActionAttributes[]; // 通貨転送アクション廃止(2024-01-29~)
-    // registerService?: IRegisterServiceAttributes[]; // サービス登録アクション廃止(2024-01-30~)
-    /**
-     * Eメール送信アクション
-     * 注文処理後のメッセージ送信として機能しているので注意
-     */
-    sendEmailMessage?: ISendEmailMessage[]; // optimize(2024-05-07~)
-}
-export interface IAttributes extends SendActionFactory.IAttributes<IObject, IResult> {
+// 注文処理中のメッセージ送信として機能しているので廃止(2024-06-21~)
+// export interface IPotentialActions {
+//     // moneyTransfer?: IMoneyTransferActionAttributes[]; // 通貨転送アクション廃止(2024-01-29~)
+//     // registerService?: IRegisterServiceAttributes[]; // サービス登録アクション廃止(2024-01-30~)
+//     // sendEmailMessage?: ISendEmailMessage[]; // optimize(2024-05-07~)
+// }
+export interface IAttributes extends Pick<
+    SendActionFactory.IAttributes<IObject, IResult>,
+    'typeOf' | 'result' | 'recipient' | 'project' | 'object' | 'error' | 'agent'
+> {
     agent: IAgent;
     recipient: IRecipient;
-    potentialActions?: IPotentialActions;
+    // potentialActions?: IPotentialActions; // discontinue(2024-06-21~)
 }
 /**
  * 注文配送アクション